welcome-clerk-13 Task: Convert to rich multi-project with 'foreign' inclusions
Pushed down aggregate component so that domain and peer artifacts can be created.
