+++ .gitignore
@@ -59,7 +59,7 @@ hs_err_pid*
 ### Gradle template
 .gradle
 **/build/
-!src/**/build/
+!domain/root/src/**/build/
 gradle-app.setting
 !/gradle/wrapper/gradle-wrapper.jar
