+++ build.gradle.kts
@@ -1,59 +1,15 @@
-import io.gitlab.arturbosch.detekt.Detekt
-import io.gitlab.arturbosch.detekt.DetektCreateBaselineTask
-
-plugins {
-    application
-    kotlin("jvm") version "1.6.21"
-    id("io.gitlab.arturbosch.detekt") version "1.20.0"
-    id("org.jetbrains.qodana") version "0.1.13"             // TBD: docker as a dependency
-}
-
 group = "me.rdd13.th15"
-version = "0.0.1"
-application {
-    mainClass.set("me.rdd13.th15.ApplicationKt")
+version = "0.1.0"
 
-    val isDevelopment: Boolean = project.ext.has("development")
-    applicationDefaultJvmArgs = listOf("-Dio.ktor.development=$isDevelopment")
-}
-
-repositories {
-    mavenCentral()
-    maven { url = uri("https://maven.pkg.jetbrains.space/public/p/ktor/eap") }
-}
-
-tasks.test {
-    useJUnitPlatform()
-}
-
-dependencies {
-    implementation("io.ktor:ktor-server-core-jvm:2.0.1")
-    implementation("io.ktor:ktor-server-netty-jvm:2.0.1")
-
-    implementation("ch.qos.logback:logback-classic:1.2.11")
-
-    testImplementation(kotlin("test"))
-    testImplementation("io.ktor:ktor-server-test-host:2.0.1")
-}
-
-// https://detekt.dev/docs/gettingstarted/gradle
-detekt {
-    buildUponDefaultConfig = true
-    allRules = false
-    config = files("$projectDir/config/detekt.yml")
-    baseline = file("$projectDir/config/baseline.xml")
-}
-
-tasks.withType<Detekt>().configureEach {
-    jvmTarget = "11"
-    reports {
-        html.required.set(true)
-        xml.required.set(true)
-        txt.required.set(true)
-        sarif.required.set(true)
+allprojects {
+    repositories {
+        google()
+        mavenCentral()
+        maven("https://maven.pkg.jetbrains.space/public/p/compose/dev")
     }
 }
-
-tasks.withType<DetektCreateBaselineTask>().configureEach {
-    jvmTarget = "11"
+plugins {
+    kotlin("multiplatform") apply false
+    id("io.gitlab.arturbosch.detekt") version "1.20.0" apply false
+    id("org.jetbrains.qodana") version "0.1.13" apply false
 }
