+++ domain/root/build.gradle.kts
@@ -1,0 +1,52 @@
+import io.gitlab.arturbosch.detekt.Detekt
+import io.gitlab.arturbosch.detekt.DetektCreateBaselineTask
+
+plugins {
+    application
+    kotlin("jvm")
+    id("io.gitlab.arturbosch.detekt")
+    id("org.jetbrains.qodana")
+}
+
+application {
+    mainClass.set("me.rdd13.th15.ApplicationKt")
+
+    val isDevelopment: Boolean = project.ext.has("development")
+    applicationDefaultJvmArgs = listOf("-Dio.ktor.development=$isDevelopment")
+}
+
+tasks.test {
+    useJUnitPlatform()
+}
+
+dependencies {
+    implementation("io.ktor:ktor-server-core-jvm:2.0.1")
+    implementation("io.ktor:ktor-server-netty-jvm:2.0.1")
+
+    implementation("ch.qos.logback:logback-classic:1.2.11")
+
+    testImplementation(kotlin("test"))
+    testImplementation("io.ktor:ktor-server-test-host:2.0.1")
+}
+
+// https://detekt.dev/docs/gettingstarted/gradle
+detekt {
+    buildUponDefaultConfig = true
+    allRules = false
+    config = files("$projectDir/config/detekt.yml")
+    baseline = file("$projectDir/config/baseline.xml")
+}
+
+tasks.withType<Detekt>().configureEach {
+    jvmTarget = "11"
+    reports {
+        html.required.set(true)
+        xml.required.set(true)
+        txt.required.set(true)
+        sarif.required.set(true)
+    }
+}
+
+tasks.withType<DetektCreateBaselineTask>().configureEach {
+    jvmTarget = "11"
+}
