+++ domain/root/config/detekt.rdd13r.yml
