+++ domain/root/config/detekt.yml
