+++ domain/root/config/license.template
