+++ domain/root/src/main/kotlin/me/rdd13/th15/Application.kt
